operators/classify_video_zero_shot/pyproject.toml
[project]
name = "feluda-classify-video-zero-shot"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
    "feluda[video]",
    "torch>=2.6.0",
    "pillow>=11.1.0",
    "transformers>=4.51.1"
    ]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.semantic_release]
version_variable = ["pyproject.toml:project.version"]

[tool.semantic_release.branches.main]
match = "main"
prerelease = false
tag_format = "{name}-{version}"

[tool.hatch.build.targets.wheel]
packages = ["."]
